Add delete vre endpoint again

+++ timbuctoo-instancev4/src/test/java/nl/knaw/huygens/timbuctoo/security/dataaccess/localfile/LocalFileVreAuthorizationAccessTest.java
@@ -1,8 +1,9 @@
 package nl.knaw.huygens.timbuctoo.security.dataaccess.localfile;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nl.knaw.huygens.timbuctoo.security.dto.VreAuthorization;
 import nl.knaw.huygens.timbuctoo.security.dataaccess.VreAuthorizationAccess;
+import nl.knaw.huygens.timbuctoo.security.dto.VreAuthorization;
+import nl.knaw.huygens.timbuctoo.security.exceptions.AuthorizationUnavailableException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -16,6 +17,7 @@ import java.util.Optional;
 import static nl.knaw.huygens.timbuctoo.security.dto.UserRoles.ADMIN_ROLE;
 import static nl.knaw.huygens.timbuctoo.security.dto.UserRoles.UNVERIFIED_USER_ROLE;
 import static nl.knaw.huygens.timbuctoo.util.OptionalPresentMatcher.present;
+import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.nullValue;
@@ -26,15 +28,20 @@ public class LocalFileVreAuthorizationAccessTest {
 
   public static final Path AUTHORIZATIONS_FOLDER = Paths.get("src", "test", "resources");
   public static final String VRE = "vre";
-  public static final Path VRE_AUTH_PATH = AUTHORIZATIONS_FOLDER.resolve(String.format("%s.json", VRE));
+  public static final String VRE_FILE = "vre.json";
+  public static final Path VRE_AUTH_PATH = AUTHORIZATIONS_FOLDER.resolve(VRE_FILE);
 
   public static final String USER_ID = "USER000000000001";
+  public static final String USER_ID_WITHOUT_WRITE_PERMISSIONS = "USER000000000002";
   private VreAuthorizationAccess instance;
   private ObjectMapper objectMapper;
 
   @Before
   public void setup() throws Exception {
-    VreAuthorization[] authorizations = {VreAuthorization.create(VRE, USER_ID, "USER")};
+    VreAuthorization[] authorizations = {
+      VreAuthorization.create(VRE, USER_ID, "USER"),
+      VreAuthorization.create(VRE, USER_ID_WITHOUT_WRITE_PERMISSIONS, UNVERIFIED_USER_ROLE)
+    };
     File file = VRE_AUTH_PATH.toFile();
     objectMapper = new ObjectMapper();
     objectMapper.writeValue(file, authorizations);
@@ -44,7 +51,9 @@ public class LocalFileVreAuthorizationAccessTest {
 
   @After
   public void teardown() throws Exception {
-    Files.delete(VRE_AUTH_PATH);
+    if (new File(VRE_AUTH_PATH.toString()).exists()) {
+      Files.delete(VRE_AUTH_PATH);
+    }
   }
 
   @Test
@@ -117,5 +126,18 @@ public class LocalFileVreAuthorizationAccessTest {
 
     // Teardown
     Files.delete(AUTHORIZATIONS_FOLDER.resolve(String.format("%s.json", newVre)));
+  }
+
+  @Test(expected = AuthorizationUnavailableException.class)
+  public void deleteVreAuthorizationThrowsAnAuthorizationUnavailableExceptionWhenTheFileIsUnavailable()
+    throws Exception {
+
+    instance.deleteVreAuthorizations("nonExisting");
+  }
+
+  @Test
+  public void deleteVreAuthorizationsDeletesTheVreAuthorizationsFile() throws Exception {
+    instance.deleteVreAuthorizations(VRE);
+    assertThat(new File(AUTHORIZATIONS_FOLDER.resolve(VRE_FILE).toString()).exists(), equalTo(false));
   }
 }
